20160422: add main

+++ src/smr-simulator.h
@@ -8,11 +8,11 @@ typedef struct
 
 typedef struct
 {
-        SSDTag    	ssd_tag;
-        int             ssd_id;			// ssd buffer location 
-        unsigned        ssd_flag;
-//	int		usage_count;
-//	int		next_freessd;
+        SSDTag     ssd_tag;
+        long       ssd_id;			// ssd buffer location 
+        unsigned   ssd_flag;
+//	long		usage_count;
+//	long		next_freessd;
 } SSDDesc;
 
 #define SSD_VALID 0x01
@@ -21,15 +21,15 @@ typedef struct
 typedef struct SSDHashBucket
 {
         SSDTag				hash_key;
-        int                             ssd_id;
+        long                             ssd_id;
         struct SSDHashBucket		*next_item;
 } SSDHashBucket;
 
 typedef struct
 {
-	int		n_usedssd;
-	int		first_usedssd;		// Head of list of used ssds
-	int		last_usedssd;		// Tail of list of used ssds
+	long		n_usedssd;
+	long		first_usedssd;		// Head of list of used ssds
+	long		last_usedssd;		// Tail of list of used ssds
 } SSDStrategyControl;
 
 extern SSDDesc		*ssd_descriptors;
@@ -43,16 +43,16 @@ extern SSDHashBucket	*ssd_hashtable;
 extern int smrread(int smr_fd, char* buffer, size_t size, off_t offset);
 extern int smrwrite(int smr_fd, char* buffer, size_t size, off_t offset);
 
-extern unsigned NSSDs;
-extern unsigned NSSDTables;
-extern unsigned SSD_SIZE;
-extern unsigned BLCKSZ;
-extern unsigned BNDSZ;
-extern unsigned INTERVALTIMELIMIT;
-extern unsigned	NSSDLIMIT;
-extern unsigned NSSDCLEAN;
+extern unsigned long NSSDs;
+extern unsigned long NSSDTables;
+extern unsigned long SSD_SIZE;
+extern size_t BLCKSZ;
+extern size_t BNDSZ;
+extern unsigned long INTERVALTIMELIMIT;
+extern unsigned	long NSSDLIMIT;
+extern unsigned long NSSDCLEAN;
 extern char     smr_device[100];
 extern char	inner_ssd_device[100];
 extern int *	inner_ssd_fd;
 extern int *	smr_fd;
-extern unsigned	interval_time;
+extern unsigned	long interval_time;
